Add files via upload
List of all the observations made by UVIT along with their SIMBAD name as well as the categorized morphology. Some instances of RA and Dec did not have any object in the 1 arcsecond radius. Hence they have been reported as 'No object found in 1 arcsecond'. It is advised that the user carry out a manual search for objects, if any, for such interactions, based on the individual need.